requirements.txt
llvmlite==0.46.0
numba==0.63.1
numpy==2.3.5
pillow==12.1.0
pygame==2.6.1
